requirements.txt
appdirs==1.4.4
black==20.8b1
click==7.1.2
mypy-extensions==0.4.3
numpy==1.19.4
pathspec==0.8.0
Pillow==8.0.1
regex==2020.10.28
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
